requirements.txt
keyboard==0.13.5
Pillow~=8.4.0
pyautogui~=0.9.53
pynput~=1.6.6
